refactor(gateway): move test utils to gateway
